refactor: remove authorization barrel export, import from ~/utils directly
The barrel re-export in .server/auth/authorization.ts added no value —
same exports, no server-specific logic. All consumers now import from
~/utils/authorization directly.
