--- app/.server/auth/authorization.ts
@@ -1,6 +1,0 @@
-export {
-  canSee,
-  canModify,
-  canCreate,
-  filterVisible,
-} from "~/utils/authorization";
